pyproject.toml
[tool.poetry]
name = "langchain agent base rag application"
version = "0.1.0"
description = ""
authors = ["Nasir Abbas <akhtarabbas.islamicreaction@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.3.3"
langchain-google-genai = "^2.0.1"
python-dotenv = "^1.0.1"
beautifulsoup4 = "^4.12.3"
faiss-cpu = "^1.9.0"
langchain-community = "^0.3.2"
requests = "^2.32.3"
tavily-python = "^0.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
